smooth: Don't crash on failure to compare floating point number values
Since we're a fan controller, we *really* don't want to panic, because
we wouldn't hand control back to the driver for fans that behave that
way (see: amdgpu). In these cases, we'd rather just assume equality for
floating point #s. The worst case is a slightly incorrect value due to a
messed up sort.